Update plugins internal data structure after `hot.updateData` method call
